+++ lib/Tetromino.h
@@ -41,7 +41,7 @@ class Tetromino{
 
 		void setPos(gf::Vector2u p);
 
-		std::set<std::pair<int, int>> getCases(); //modifier avec Vector2i
+		std::set<std::pair<uint8_t, uint8_t>> getCases(); //modifier avec Vector2i
 
 		void rotate();
